+++ src/google_music_scripts/commands/sync.py
@@ -234,18 +234,6 @@ def sync_down(
 	help="Delete successfully uploaded local files."
 )
 @click.option(
-	'--transcode-lossless/--no-transcode-lossless',
-	is_flag=True,
-	default=True,
-	help="Transcode lossless files to MP3 for upload."
-)
-@click.option(
-	'--transcode-lossy/--no-transcode-lossy',
-	is_flag=True,
-	default=True,
-	help="Transcode non-MP3 lossy files to MP3 for upload."
-)
-@click.option(
 	'-f', '--filters',
 	metavar='FILTER',
 	multiple=True,
@@ -268,8 +256,6 @@ def sync_up(
 	no_recursion,
 	max_depth,
 	delete_on_success,
-	transcode_lossless,
-	transcode_lossy,
 	filters,
 	input_paths
 ):
@@ -313,8 +299,6 @@ def sync_up(
 		upload_songs(
 			mm,
 			to_upload,
-			transcode_lossless=transcode_lossless,
-			transcode_lossy=transcode_lossy,
 			delete_on_success=delete_on_success
 		)